pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zmq-pipeline"
version = "0.1.0"
description = "Distributed ZMQ Ventilator/Worker/Sink pipeline framework"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "pyzmq>=25.0",
    "fastapi>=0.135",
    "uvicorn>=0.41",
    "numpy>=1.24",
    "pydantic>=2.0",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
    "typing_extensions>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools]
# This tells setuptools to look for your code specifically in the src directory
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"